% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Find the amount of data within x standard deviations of the mean}
\usage{
chebyshev(x)
}
\arguments{
\item{x}{A number}
}
\value{
A number
}
\description{
Find the amount of data within x standard deviations of the mean
}
\examples{
chebyshev(2)
}
