% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRING234753bart0074.R
\docType{package}
\name{SPRING234753bart0074}
\alias{SPRING234753bart0074}
\title{Math 4073 Package}
\description{
Contains utility functions for Applied Statistical Methods
}
\author{
Michael Bartlett \email{michael@ou.edu}
}
